refactor(devtool): move code around, organize
