+++ devtool/package.json
@@ -22,7 +22,7 @@
     "prelib": "rm -rf lib/ && mkdir -p lib && ../node_modules/.bin/typings install",
     "lib": "../node_modules/.bin/tsc",
     "dist-graphSerializer": "../node_modules/.bin/browserify lib/graphSerializer.js -s gs --outfile dist/graphSerializer.js",
-    "dist-panel": "../node_modules/.bin/browserify lib/panel.js -s ppp --outfile dist/panel.js",
+    "dist-panel": "../node_modules/.bin/browserify lib/panel/index.js -s ppp --outfile dist/panel.js",
     "copy-ready-src": "cp lib/background.js lib/contentScript.js lib/launcher.js src/launcher.html src/manifest.json src/mermaid.css src/panel.html src/128.png dist/",
     "zip": "zip cyclejs-devtool.zip dist/*",
     "dist": "mkdir -p dist && npm install && npm run lib && npm run dist-graphSerializer && npm run dist-panel && npm run copy-ready-src && npm run zip",
